useSqlBuilder removal cleanup
