+++ web/renderer/components/DataTable/index.tsx
@@ -7,7 +7,7 @@ import {
   RowForDataTableFragment,
 } from "@gen/graphql-types";
 import DataTableLayout from "@layouts/DataTableLayout";
-import { RefParams, TableParams } from "@lib/params";
+import { TableParams } from "@lib/params";
 import { ReactNode } from "react";
 import AddRowsButton from "./AddRowsButton";
 import ShowAllColumns from "./ShowAllColumns";
@@ -21,7 +21,6 @@ type Props = {
   rows?: RowForDataTableFragment[];
   columns?: ColumnForDataTableFragment[];
   message?: ReactNode | null;
-  params: RefParams & { tableName?: Maybe<string>; q: string };
   error?: ApolloError;
   warnings?: Maybe<string[]>;
 };
@@ -53,11 +52,7 @@ export function Inner({ columns, rows, message = null, ...props }: Props) {
 
 type DataTableParams = TableParams & { offset?: Maybe<number> };
 
-type TableProps = {
-  params: DataTableParams & { q: string };
-};
-
-function WithContext(props: TableProps) {
+function WithContext() {
   const {
     loading,
     loadingWorkingDiff,
@@ -84,7 +79,6 @@ function WithContext(props: TableProps) {
 
   return (
     <Inner
-      params={props.params}
       loadMore={workingDiffRowsToggled ? loadMoreWorkingDiff : loadMore}
       rows={workingDiffRowsToggled ? workingDiffRows : rows}
       columns={columns}
@@ -99,7 +93,7 @@ export default function DataTable(props: { params: DataTableParams }) {
   return (
     <>
       <DataTableLayout params={params} tableName={props.params.tableName}>
-        <WithContext params={params} />
+        <WithContext />
       </DataTableLayout>
       <AddRowsButton {...props} />
     </>
